updated command line
